fix menu bugs
