+++ app/js/system/MainMenu.js
@@ -5,8 +5,8 @@
  */
 
 const { remote } = require('electron')
+
 const { Menu } = remote
-const menuTemplate = require('./menuTemplate')
 
 class MainMenu {
 
@@ -16,20 +16,149 @@ class MainMenu {
 
   constructor(app) {
     this.app = app
+    const template = this.createTemplate()
 
     // Add theme sub menu
-    menuTemplate[2].submenu[7].submenu = this.createThemeSubMenu()
+    template[2].submenu[7].submenu = this.createThemeSubMenu()
 
     // Finally, build menu
-    Menu.setApplicationMenu(Menu.buildFromTemplate(menuTemplate))
+    Menu.setApplicationMenu(Menu.buildFromTemplate(template))
   }
 
   /**
    *  Callback for when theme button is clicked.
    */
 
-  themeClicked(menuItem, browserWindow, event) {
+  themeClicked(menuItem) {
     this.app.setTheme(menuItem.theme)
+  }
+
+  createTemplate() {
+    const template = [{
+      label: 'Edit',
+      submenu: [{
+        role: 'undo'
+      }, {
+        role: 'redo'
+      }, {
+        type: 'separator'
+      }, {
+        role: 'cut'
+      }, {
+        role: 'copy'
+      }, {
+        role: 'paste'
+      }, {
+        role: 'selectall'
+      }, {
+        type: 'separator'
+      }, {
+        label: 'Format',
+        click: () => {
+          this.app.getCurrentPage().editor.format()
+        }
+      }, {
+        label: 'Minify',
+        click: () => {
+          this.app.getCurrentPage().editor.minify()
+        }
+      }]
+    }, {
+      label: 'View',
+      submenu: [{
+        role: 'reload'
+      }, {
+        role: 'toggledevtools'
+      }, {
+        type: 'separator'
+      }, {
+        role: 'resetzoom'
+      }, {
+        role: 'zoomin'
+      }, {
+        role: 'zoomout'
+      }, {
+        type: 'separator'
+      }, {
+        label: 'Theme',
+        submenu: []
+      }, {
+        type: 'separator'
+      }, {
+        role: 'togglefullscreen'
+      }]
+    }, {
+      role: 'window',
+      submenu: [{
+        role: 'minimize'
+      }, {
+        role: 'close'
+      }]
+    }, {
+      role: 'help',
+      submenu: [{
+        label: 'Learn More',
+        click: () => {
+          electron.shell.openExternal('http://electron.atom.io')
+        }
+      }]
+    }]
+    if (process.platform === 'darwin') {
+      template.unshift({
+          label: 'JSON-Splora',
+          submenu: [{
+            role: 'about'
+          }, {
+            type: 'separator'
+          }, {
+            role: 'services',
+            submenu: []
+          }, {
+            type: 'separator'
+          }, {
+            role: 'hide'
+          }, {
+            role: 'hideothers'
+          }, {
+            role: 'unhide'
+          }, {
+            type: 'separator'
+          }, {
+            role: 'quit'
+          }]
+        })
+        // Edit menu.
+      template[1].submenu.push({
+          type: 'separator'
+        }, {
+          label: 'Speech',
+          submenu: [{
+            role: 'startspeaking'
+          }, {
+            role: 'stopspeaking'
+          }]
+        })
+        // Window menu.
+      template[3].submenu = [{
+        label: 'Close',
+        accelerator: 'CmdOrCtrl+W',
+        role: 'close'
+      }, {
+        label: 'Minimize',
+        accelerator: 'CmdOrCtrl+M',
+        role: 'minimize'
+      }, {
+        label: 'Zoom',
+        role: 'zoom'
+      }, {
+        type: 'separator'
+      }, {
+        label: 'Bring All to Front',
+        role: 'front'
+      }]
+    }
+
+    return template
   }
 
   /**
@@ -41,36 +170,36 @@ class MainMenu {
       label: 'Default',
       theme: 'default',
       type: 'radio',
-      click: (menuItem, browserWindow, event) => {
-        this.themeClicked(menuItem, browserWindow, event)
+      click: (menuItem) => {
+        this.themeClicked(menuItem)
       }
     }, {
       label: 'Elegant',
       theme: 'Elegant',
       type: 'radio',
-      click: (menuItem, browserWindow, event) => {
-        this.themeClicked(menuItem, browserWindow, event)
+      click: (menuItem) => {
+        this.themeClicked(menuItem)
       }
     }, {
       label: 'Mdn-Like',
       theme: 'mdn-like',
       type: 'radio',
-      click: (menuItem, browserWindow, event) => {
-        this.themeClicked(menuItem, browserWindow, event)
+      click: (menuItem) => {
+        this.themeClicked(menuItem)
       }
     }, {
       label: 'Neat',
       theme: 'neat',
       type: 'radio',
-      click: (menuItem, browserWindow, event) => {
-        self.themeClicked(menuItem, browserWindow, event)
+      click: (menuItem) => {
+        self.themeClicked(menuItem)
       }
     }, {
       label: 'Neo',
       theme: 'neo',
       type: 'radio',
-      click: (menuItem, browserWindow, event) => {
-        this.themeClicked(menuItem, browserWindow, event)
+      click: (menuItem) => {
+        this.themeClicked(menuItem)
       }
     }]
   }
