+++ package.json
@@ -7,7 +7,7 @@
     "jsplora": "bin/jsplora"
   },
   "scripts": {
-    "validate": "node_modules/.bin/eslint app/js -c .eslintrc.json",
+    "validate": "node_modules/.bin/eslint index.js app/js -c .eslintrc.json",
     "start": "ENV=development electron .",
     "build-darwin": "node_modules/.bin/electron-packager . --darwin --overwrite",
     "build-linux": "node_modules/.bin/electron-packager . --linux --overwrite",
